TravelTalesBackend/requirements.txt
requests
scikit-learn
python-dotenv
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic